Separate functions according to their interests
