+++ src/home-page/page-display.js
@@ -1,57 +1,4 @@
-import { lookupShow, getComments } from '../api/api.js';
-
-const addItemToList = (ul, arr) => {
-  arr.forEach((el) => {
-    ul.innerHTML += `<li class="comment-list-item my-1">${el.creation_date} ${el.username}: ${el.comment}</li>`;
-  });
-};
-
-async function gotoCommentPage(showId) {
-  const showInfo = await lookupShow(showId);
-  const commentsInfo = await getComments(showId);
-  const modal = document.querySelector('.modal');
-  modal.classList.add('active');
-
-  modal.innerHTML = `
-  <div
-          class="
-            modal-content
-            d-flex
-            flex-column
-            align-items-center
-            justify-content-center
-          "
-        >
-          <div class="d-flex flex-row w-100 justify-content-center">
-            <img
-              src="${showInfo.image.medium}"
-              alt="shows"
-              class="popup-image"
-            />
-            <div class="close">&times;</div>
-          </div>
-          <h3 class="popup-title my-4">${showInfo.name}</h3>
-          <div class="info d-flex flex-row my-2 w-75">
-            <div class="categories d-flex flex-column align-items-start w-50">
-              <h5>Genre : ${showInfo.genres[0]}</h5>
-              <h5>Language : ${showInfo.language}</h5>
-            </div>
-            <div class="categories d-flex flex-column align-items-start w-50">
-              <h5>Release Date : ${showInfo.premiered.split('-')[0]}</h5>
-              <h5>Rating : ${showInfo.rating.average}</h5>
-            </div>
-          </div>
-          <h4 class="comments-title my-2">Comments</h4>
-          <ul class="comments-list list-group list-unstyled">
-          </ul>
-        </div>
-  `;
-  const ul = document.querySelector('.comments-list');
-  addItemToList(ul, commentsInfo);
-
-  const closeButton = document.querySelector('.close');
-  closeButton.addEventListener('click', () => modal.classList.remove('active'));
-}
+import gotoCommentPage from '../comments/popup.js';
 
 class ShowContainer {
   constructor(showImage, showName, showLikes, showId) {
@@ -111,4 +58,4 @@ const displayShows = (shows) => {
   });
 };
 
-export { displayShows, addItemToList };
+export default displayShows;
